FEWD Final Project Submission
What is your site? Why did you want to create this project? Explain your inspiration and help sell the idea of your website.
‘Transform your patient experience’ is the key idea for this website - a lead generation microsite for the company I work at - Outcome Health. The idea is to show the true value of benefits of the company’s products (versus just outlining what the products are).

What was the hardest part of this project for you? Describe the largest challenge you encountered
The carousel slideshow really set me back. It wasn’t working the way I wanted to and in the future I also might want to switch it to a different type of carousel.

After completion, what are you most proud of?
Figuring out mobile responsive. As a designer primarily in print, this is probably one of the biggest differences in thinking about layout so this was a great exercise in understanding best practices for mobile design.

What’s going to be in version 2.0? What will you add next?
Version 2.0 will be finishing up the home page, adding 2-3 additional pages, and making it functional for lead gen purposes (i.e. forms).
I am also working on troubleshooting the mobile version to fix a width issue, and look into the hover function (currently photos that are linked click through without the user being able to read what the hover is).